+++ jdaviz/configs/imviz/plugins/coords_info/coords_info.vue
@@ -13,13 +13,13 @@
         </tr>
         <tr :style="unreliable_world ? 'color: #B8B8B8' : ''">
           <td width="42"><b>{{ world_label_prefix }}</b></td>
-          <td width="115">{{ world_ra }}</td>
+          <td :width="world_label_prefix == 'Wave' ? 85 : 115">{{ world_ra }}</td>
           <td width="120">{{ world_dec }}</td>
           <td>{{ world_label_icrs }}</td>
         </tr>
         <tr :style="unreliable_world ? 'color: #B8B8B8' : ''">
-          <td width="42">{{ unreliable_world ? '(est.)' : '' }}</td>
-          <td width="115">{{ world_ra_deg }}</td>
+          <td width="42" :style="unreliable_world ? '' : 'font-weight: bold'">{{ world_label_prefix_2 }}</td>
+          <td :width="world_label_prefix == 'Wave' ? 85 : 115">{{ world_ra_deg }}</td>
           <td width="120">{{ world_dec_deg }}</td>
           <td>{{ world_label_deg }}</td>
         </tr>