+++ CHANGES.rst
@@ -114,6 +114,8 @@ Cubeviz
 
 - Added Slice plugin player control buttons. [#1848]
 
+- Improved mouseover info display for spectrum viewer. [#1894]
+
 Imviz
 ^^^^^
 
@@ -134,12 +136,16 @@ Mosviz
 
 - ``load_data`` method can now load JWST NIRCam and NIRSpec level 2 data. [#1835]
 
+- Improved mouseover info display for spectrum viewer. [#1894]
+
 Specviz
 ^^^^^^^
 
 - Spectrum viewer now shows X and Y values under cursor. [#1759]
 
 - Switch to opt-in concatenation for multi-order x1d spectra. [#1659]
+
+- Improved mouseover info display for spectrum viewer. [#1894]
 
 Specviz2d
 ^^^^^^^^^
@@ -148,6 +154,8 @@ Specviz2d
   with Polynomial, Spline, and Legendre options. [#1889]
 
 - Add dropdown for choosing background statistic (average or median). [#1922]
+
+- Improved mouseover info display for spectrum viewer. [#1894]
 
 API Changes
 -----------